Move variable/identifier family out-of-line into MLIRGenVariables.cpp (review doc 1)
20 methods (variable registration incl. this-context and declaration,
array/object binding patterns with sub-paths and spread, declaration
name/item processing, VariableDeclaration/List, parameter bindings,
and the identifier/type/namespace resolution family) move from
MLIRGenImpl.h into a new MLIRGenVariables.cpp TU; header keeps
declarations with defaults. 1223 lines out of the header
(11902 -> 10679).
